+++ front/src/app/(auth)/components/FormField.tsx
@@ -1,0 +1,36 @@
+import { UseFormRegister, FieldValues, Path } from "react-hook-form";
+
+interface FormFieldProps<T extends FieldValues> {
+  label: string;
+  name: Path<T>;
+  type: string;
+  register: UseFormRegister<T>;
+  error?: string;
+}
+
+export function FormField<T extends FieldValues>({
+  label,
+  name,
+  type,
+  register,
+  error,
+}: FormFieldProps<T>) {
+  return (
+    <div className="mb-4">
+      <label
+        className="block text-gray-700 text-sm font-bold mb-2"
+        htmlFor={name}
+      >
+        {label}
+      </label>
+      <input
+        {...register(name)}
+        className="shadow appearance-none border-2 border-gray-300 rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+        id={name}
+        type={type}
+        placeholder={label}
+      />
+      {error && <p className="text-red-500 text-xs italic mt-1">{error}</p>}
+    </div>
+  );
+}
