+++ front/src/app/(auth)/components/toast.tsx
